Extracted #to_s related methods to separate module
